Software/requirements.txt
fastapi
uvicorn
torch
torchvision
pillow
python-multipart
mysql-connector-python